dd ensureNeverFrozen
