+++ apollo-normalized-cache/src/commonMain/kotlin/com/apollographql/apollo3/cache/normalized/Platform.kt
@@ -1,0 +1,7 @@
+package com.apollographql.apollo3.cache.normalized
+
+expect object Platform {
+  fun currentTimeMillis(): Long
+  fun currentThreadId(): String
+  fun ensureNeverFrozen(obj: Any)
+}
