--- apollo-normalized-cache/src/commonMain/kotlin/com/apollographql/apollo3/cache/normalized/internal/Platform.kt
@@ -1,5 +1,0 @@
-package com.apollographql.apollo3.cache.normalized.internal
-
-internal expect object Platform {
-  fun currentTimeMillis(): Long
-}
