+++ apollo-normalized-cache/src/jsMain/kotlin/com/apollographql/apollo3/cache/normalized/Platform.kt
@@ -1,0 +1,16 @@
+package com.apollographql.apollo3.cache.normalized
+
+import kotlin.js.Date
+
+actual object Platform {
+  actual fun currentTimeMillis(): Long {
+    return Date().getTime().toLong()
+  }
+
+  actual fun currentThreadId(): String {
+    return "js"
+  }
+
+  actual fun ensureNeverFrozen(obj: Any) {
+  }
+}
