--- apollo-normalized-cache/src/jsMain/kotlin/com/apollographql/apollo3/cache/normalized/internal/Platform.kt
@@ -1,10 +1,0 @@
-package com.apollographql.apollo3.cache.normalized.internal
-
-import kotlin.js.Date
-
-internal actual object Platform {
-  actual fun currentTimeMillis(): Long {
-    return Date().getTime().toLong()
-  }
-}
-
