+++ apollo-normalized-cache/src/jvmMain/kotlin/com/apollographql/apollo3/cache/normalized/Platform.kt
@@ -1,0 +1,14 @@
+package com.apollographql.apollo3.cache.normalized
+
+ actual object Platform {
+  actual fun currentTimeMillis(): Long {
+    return System.currentTimeMillis()
+  }
+
+  actual fun currentThreadId(): String {
+    return Thread.currentThread().id.toString()
+  }
+
+   actual fun ensureNeverFrozen(obj: Any) {
+   }
+ }
