--- apollo-normalized-cache/src/jvmMain/kotlin/com/apollographql/apollo3/cache/normalized/internal/Platform.kt
@@ -1,7 +1,0 @@
-package com.apollographql.apollo3.cache.normalized.internal
-
-internal actual object Platform {
-  actual fun currentTimeMillis(): Long {
-    return System.currentTimeMillis()
-  }
-}
