+++ composite/integration-tests-kotlin/src/appleTest/kotlin/com/apollographql/apollo3/integration/Platform.kt
@@ -38,6 +38,3 @@ actual fun checkTestFixture(actualText: String, name: String) {
   assertEquals(actualText, readTestFixture(name))
 }
 
-actual fun currentThreadId(): String {
-  return pthread_self()?.toString() ?: "?"
-}
