+++ composite/integration-tests-kotlin/src/jvmTest/kotlin/com/apollographql/apollo3/integration/Platform.kt
@@ -32,7 +32,3 @@ actual fun checkTestFixture(actualText: String, name: String) {
     }
   }
 }
-
-actual fun currentThreadId(): String {
-  return Thread.currentThread().id.toString()
-}
